cleaned up docstrings and added more common code to master_model.py
